requirements.txt
aiohttp==3.8.1
pandas==1.3.5
python-dotenv==0.19.2
